+++ paper/paper.bib
@@ -12,7 +12,6 @@
   month = mar,
   year = {2018},
   keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
-  file = {/home/steven/Dropbox/PaperCatalogueZotero/arXiv1803.02477 [astro-ph]/2018/Wolz et al_2018_Intensity mapping cross-correlations II.pdf;/home/steven/Zotero/storage/32FN5HG5/1803.html},
   annote = {Comment: 15 pages, 8 figures, 3 tables. Comments welcome}
 }
 
