### 2018-07-24 b8f36f0
remove redundant github link
