+++ paper/paper.bib
@@ -76,7 +76,7 @@
 
 
 @online{github,
-  author = {Steven G. Murray},
+  author = {S. G. Murray},
   title = {powerbox: Make arbitrarily structured, atbitrary-dimension boxes and log-normal mocks},
   year = 2018,
   url = {https://github.com/steven-murray/powerbox},
