### 2018-07-26 #850 @whedon
[REVIEW]: powerbox: A Python package for creating structured fields with isotropic power spectra
**Submitting author:** @steven-murray (<a href="http://orcid.org/0000-0003-3059-3823">Steven Murray</a>)
**Repository:** <a href="https://github.com/steven-murray/powerbox" target ="_blank">https://github.com/steven-murray/powerbox</a>
**Version:** v0.5.5
