### 2018-08-14 #850 @dfm
I'm starting to go through this and here are a few comments as I go:

- The travis build is currently failing with some sort of coverage bug.
- I think that the quickstart tutorial could have a few more words about the details of what is meant by "power spectrum" (e.g. units, normalization, etc.) and what the properties are required by the function provided to the PowerBox class.
- The powerbox version in the [quickstart tutorial](http://powerbox.readthedocs.io/en/latest/demos/getting_started.html) is not the same as the one that it is documenting (v0.5.4 vs 0.5.5).
- The arguments to `create_discrete_sample` are not all documented. (I was not clear on what `min_at_zero` meant until I looked at the source)
- I think that it would also be useful to have a discussion of the specific numerical method (perhaps in the context of the relevant literature) somewhere in the docs. For example, how does this method compare to a Gaussian Process method that wouldn't require discretization for sampling. This might also lead to a useful discussion of how to choose N for specific use cases.
- The docs need a discussion of the dependencies and how to contribute.
