### 2018-08-19 59980d8
attempt to resolve travis build error
