### 2018-08-20 #850 @steven-murray
> The travis build is currently failing with some sort of coverage bug.

This is fixed by https://github.com/steven-murray/powerbox/commit/59980d84365099fab31e56238dfeb4ec7d3c2cfc
