### 2018-08-20 50efcde
updated demo versions to 0.5.5
