### 2018-08-20 #850 @steven-murray
> The powerbox version in the quickstart tutorial is not the same as the one that it is documenting (v0.5.4 vs 0.5.5).

Fixed in https://github.com/steven-murray/powerbox/commit/50efcded235e8d046c4ef28447e41019b017b307
