### 2018-08-20 4e7cdcb
improved docstring for create_discrete_sample
