+++ powerbox/powerbox.py
@@ -223,7 +223,7 @@ class PowerBox(object):
         return dk
 
     def create_discrete_sample(self, nbar, randomise_in_cell=True, min_at_zero=False,
-                               store_pos=False, seed=None):
+                               store_pos=False):
         r"""
         Assuming that the real-space signal represents an over-density with respect to some mean, create a sample
         of tracers of the underlying density distribution.
@@ -232,10 +232,21 @@ class PowerBox(object):
         ----------
         nbar : float
             Mean tracer density within the box.
-        """
-        if seed:
-            np.random.seed(seed)
+        randomise_in_cell : bool, optional
+            Whether to randomise the positions of the tracers within the cells, or put them at the grid-points (more
+            efficient).
+        min_at_zero : bool, optional
+            Whether to make the lower corner of the box at the origin, otherwise the centre of the box is at the
+            origin.
+        store_pos : bool, optional
+            Whether to store the sample of tracers as an instance variable `tracer_positions`.
 
+        Returns
+        -------
+        tracer_positions : float, array_like
+            ``(n, d)``-array, with ``n`` the number of tracers and ``d`` the number of dimensions. Each row represents
+            a single tracer's co-ordinates.
+        """
         dx = self.delta_x()
         dx = (dx + 1)*self.dx ** self.dim*nbar
         n = dx
