### 2018-08-20 #850 @steven-murray
> The arguments to create_discrete_sample are not all documented. (I was not clear on what min_at_zero meant until I looked at the source)

Fixed in https://github.com/steven-murray/powerbox/commit/4e7cdcb7bb36b9be4149adf9bcd4bda0a1ea4751
