### 2018-08-20 e49a260
added contribution guidelines and dependency discussion
