+++ CONTRIBUTORS.rst
@@ -1,7 +1,8 @@
-Author
-------
+Authors
+-------
+
 * `Steven Murray <https://github.com/steven-murray>`_
 
 Comments, corrections and suggestions
--------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * `Chris Jordan <https://github.com/cjordan>`_
