### 2018-08-20 #850 @steven-murray
> The docs need a discussion of the dependencies and how to contribute.

Added some of these in https://github.com/steven-murray/powerbox/commit/e49a26037b69c9c29f4d1006be68fe06d8a36c3f. Please let me know if they are not adequate.
