### 2018-08-20 04e6a6d
added intro to power spectrum, units, conventions etc
