### 2018-08-20 #850 @steven-murray
> I think that the quickstart tutorial could have a few more words about the details of what is meant by "power spectrum" (e.g. units, normalization, etc.) and what the properties are required by the function provided to the PowerBox class.

I've had a go of this in https://github.com/steven-murray/powerbox/commit/04e6a6de0f0c969e933affe1402b9a9ec45a833c.

The discussion is in the quickstart tutorial introduction. Please let me know if this is not quite what you were thinking and can be improved.
