### 2018-08-20 086cb71
added DOIs to some references
