+++ paper/paper.bib
@@ -58,6 +58,7 @@
 @article{Coles1991,
   title = {A Lognormal Model for the Cosmological Mass Distribution},
   volume = {248},
+  doi = {10.1093/mnras/248.1.1},
   abstract = {SUMMARY We discuss the use of a lognormal (LN) random field as a model for the distribution of matter in the Universe. We find a number of reasons why this should be a plausible approximation to the distribution of density irregularities obtained by evolving from Gaussian initial conditions. Unlike straightforward linear theory, the model always has p$>$0 but is arbitrarily close to the Gaussian at early times. It has the added advantage that, like the Gaussian model, all its statistical properties can be formulated in terms of one covariance function. A number of interesting and important difficulties with the statistical treatment of density perturbations are revealed by an analysis of this model. In particular, the LN model is not completely specified by its moments. We explain why this could be true for the actual matter field. We also show that the usual method of representing the three-and four-point correlation functions of galaxies, in terms of the parameters Q and R, is not useful for discriminating between Gaussian and non-Gaussian fluctua-tions, and propose better parameterizations in terms of the skewness and kurtosis of the three-and four-point distributions, respectively. Other characteristics of the model, such as topology (genus curves, etc.), multi-fractal behaviour, void probabilities and biasing (behaviour of 'peaks' relative to background fluctuations) are also discussed. The model also provides a way of check-ing the consistency of treatments of large-scale streaming motions in the Universe by allowing us to determine the scale at which linear theory cannot be accurate for both the matter and velocity fields. We discuss a possible model for the number-count distribution of galaxies, based on the LN distribution but allowing for discreteness effects which can make the distribution of log n appear non-Gaussian, and show how to construct Monte-Carlo simulations of point patterns (in one-, two, or three-dimensions) which contain correlations of all orders.},
   journal = {Mon. Not. R. Astron. Soc.},
   author = {Coles, Peter and Jones, Bernard},
@@ -66,12 +67,15 @@
   file = {/home/steven/Dropbox/PaperCatalogueZotero/Mon. Not. R. astr. Soc/1991/Coles_Jones_1991_A lognormal model for the cosmological mass distribution.pdf}
 }
 
-@book{Peacock1999,
-  address = {Cambridge},
-  title = {Cosmological {{Physics}}},
-  publisher = {{Cambridge University Press}},
-  author = {Peacock, John},
-  year = {1999}
+@inbook{Peacock1999,
+  place={Cambridge},
+  title={Cosmological density fields},
+  DOI={10.1017/CBO9780511804533.017},
+  booktitle={Cosmological Physics},
+  publisher={Cambridge University Press},
+  author={Peacock, J. A.},
+  year={1998},
+  pages={495–552}
 }
 
 
