### 2018-08-20 #850 @steven-murray
> a few of the references in the paper are missing DOIs

I have added DOIs to Coles 1991 and Peacock 1999 (cf https://github.com/steven-murray/powerbox/commit/086cb71443cdcfbb0ce6f08a6ac08df72898c49a). The book "Statistical Fluid Mechanics" by Monin and Yaglom does not seem to have a DOI. The article by Wolz2018 does not yet have a DOI as it is not yet accepted.
