### 2018-08-20 6078a42
added discussion of algorithm
