+++ docs/examples.rst
@@ -8,5 +8,6 @@ Other examples can be found in the :doc:`API documentation <api>` for each objec
    :maxdepth: 2
 
    demos/getting_started
+   demos/algorithm
    demos/cosmological_fields
    demos/dft
