### 2018-08-20 #850 @steven-murray
> I think that it would also be useful to have a discussion of the specific numerical method (perhaps in the context of the relevant literature) somewhere in the docs. For example, how does this method compare to a Gaussian Process method that wouldn't require discretization for sampling. This might also lead to a useful discussion of how to choose N for specific use cases.

In https://github.com/steven-murray/powerbox/commit/6078a428012ac6e8e496d8c77dae5baf680464f9 I have added a new example notebook discussing the algorithm in a little more depth. I am not quite sure how to relate this method to Gaussian Processes, so I have left this out for now. I have commented that no information on the power is available for scales less than L/N, and given a rule-of-thumb for choosing N.

Thank you for all your comments so far. I hope I have addressed them reasonably!
