### 2018-08-20 #850 @dfm
Good stuff! I'm happy with these changes and everything looks good to me now.
@arfon: I'm happy to recommend this for acceptance.
@steven-murray: congrats on the excellent software and documentation!
