### 2018-08-20 #850 @steven-murray
@dfm : Thanks for the smooth and thorough review process!
@arfon : Thanks, the DOI is 10.5281/zenodo.1400822 .

I will definitely be publishing with JOSS again!
